+++ src/test/java/net/joelinn/quartz/RedisJobStoreTest.java
@@ -1,32 +1,15 @@
 package net.joelinn.quartz;
 
-import com.google.common.base.Joiner;
 import net.joelinn.quartz.jobstore.RedisJobStore;
-import net.joelinn.quartz.jobstore.RedisJobStoreSchema;
 import org.junit.Test;
 import org.quartz.JobDetail;
-import org.quartz.Scheduler;
-import org.quartz.SchedulerException;
 import org.quartz.Trigger;
-import org.quartz.impl.StdSchedulerFactory;
-import org.quartz.impl.triggers.CronTriggerImpl;
-import org.quartz.spi.JobStore;
-import org.quartz.spi.SchedulerSignaler;
-import redis.clients.jedis.*;
-import redis.clients.util.Pool;
-import redis.embedded.RedisCluster;
-import redis.embedded.util.JedisUtil;
 
 import java.util.Map;
 import java.util.Properties;
 import java.util.Set;
 
-import static org.hamcrest.CoreMatchers.not;
-import static org.hamcrest.CoreMatchers.nullValue;
-import static org.hamcrest.MatcherAssert.assertThat;
-import static org.hamcrest.collection.IsMapContaining.hasKey;
 import static org.junit.Assert.assertEquals;
-import static org.mockito.Mockito.mock;
 
 /**
  * Joe Linn